Update Coins.java
added a for loop that I think will work to fill the array with the proper coin values. I haven't actually tested it so it is commented out. Once that array is built properly then just need a method that reads that array when searching for the coinChange needed. That way we are not building the array each time.